Backport initApplication, and fix tests
