+++ test/unit/modules/blocks/process.js
@@ -3,12 +3,12 @@
 var expect = require('chai').expect;
 var async = require('async');
 
+var application = require('../../../common/application.js');
 var node = require('../../../node');
 var modulesLoader = require('../../../common/modulesLoader');
 var genesisBlock = require('../../../../genesisBlock.json');
 var loadTables = require('./processTablesData.json');
-var clearDatabaseTable = require('../../../common/globalBefore').clearDatabaseTable;
-var DBSandbox = require('../../../common/globalBefore').DBSandbox;
+var clearDatabaseTable = require('../../../common/DBSandbox.js').clearDatabaseTable;
 
 describe('blocks/process', function () {
 
@@ -18,48 +18,28 @@ describe('blocks/process', function () {
 	var blocksVerify;
 	var accounts;
 	var db;
-	var dbSandbox;
 	var scope;
 	var originalBlockRewardsOffset;
 
 	before(function (done) {
-		dbSandbox = new DBSandbox(modulesLoader.scope.config.db, 'lisk_test_blocks_process');
-		dbSandbox.create(function (err, __db) {
-			modulesLoader.db = __db;
-			db = __db;
-			// Force rewards start at 150-th block
-			originalBlockRewardsOffset = node.constants.rewards.offset;
-			node.constants.rewards.offset = 150;
-			node.initApplication(function (err, __scope) {
-				scope = __scope;
-				accounts = __scope.modules.accounts;
-				blocksProcess = __scope.modules.blocks.process;
-				blocksVerify = __scope.modules.blocks.verify;
-				blockLogic = __scope.logic.block;
-				blocks = __scope.modules.blocks;
-				db = __scope.db;
-				done(err);
-			}, {db: db});
+		// Force rewards start at 150-th block
+		originalBlockRewardsOffset = node.constants.rewards.offset;
+		node.constants.rewards.offset = 150;
+		application.init({sandbox: {name: 'lisk_test_blocks_process'}}, function (err, scope) {
+			scope = scope;
+			accounts = scope.modules.accounts;
+			blocksProcess = scope.modules.blocks.process;
+			blocksVerify = scope.modules.blocks.verify;
+			blockLogic = scope.logic.block;
+			blocks = scope.modules.blocks;
+			db = scope.db;
+			done(err);
 		});
 	});
 
 	after(function (done) {
-		async.every([
-			'blocks where height > 1',
-			'trs where "blockId" != \'6524861224470851795\'',
-			'mem_accounts where address in (\'2737453412992791987L\', \'2896019180726908125L\')',
-			'forks_stat',
-			'votes where "transactionId" = \'17502993173215211070\''
-		], function (table, seriesCb) {
-			clearDatabaseTable(db, modulesLoader.logger, table, seriesCb);
-		}, function (err) {
-			if (err) {
-				done(err);
-			}
-			node.constants.rewards.offset = originalBlockRewardsOffset;
-			dbSandbox.destroy(modulesLoader.logger);
-			node.appCleanup(done);
-		});
+		node.constants.rewards.offset = originalBlockRewardsOffset;
+		application.cleanup(done);
 	});
 
 	beforeEach(function (done) {
