+++ test/unit/modules/blocks/verify.js
@@ -5,10 +5,10 @@ var async = require('async');
 var sinon = require('sinon');
 
 var node = require('../../../node');
+var application = require('../../../common/application.js');
 var modulesLoader = require('../../../common/modulesLoader');
 var exceptions = require('../../../../helpers/exceptions.js');
-var clearDatabaseTable = require('../../../common/globalBefore').clearDatabaseTable;
-var DBSandbox = require('../../../common/globalBefore').DBSandbox;
+var clearDatabaseTable = require('../../../common/DBSandbox.js').clearDatabaseTable;
 
 var crypto = require('crypto');
 var bson = require('../../../../helpers/bson.js');
@@ -217,25 +217,18 @@ describe('blocks/verify', function () {
 	var blocks;
 	var blockLogic;
 	var delegates;
-
 	var db;
-	var dbSandbox;
 
 	before(function (done) {
-		dbSandbox = new DBSandbox(modulesLoader.scope.config.db, 'lisk_test_blocks_verify');
-		dbSandbox.create(function (err, __db) {
-			modulesLoader.db = __db;
-			db = __db;
-			done(err);
-		});
-	});
-
-	after(function () {
-		dbSandbox.destroy(modulesLoader.logger);
-	});
-
-	before(function (done) {
-		node.initApplication(function (err, scope) {
+		application.init({
+			sandbox: {
+				name: 'lisk_test_blocks_verify'
+			},
+			scope: {
+				bus: modulesLoader.scope.bus
+			},
+			waitForGenesisBlock: false
+		}, function (err, scope) {
 			scope.modules.blocks.verify.onBind(scope.modules);
 			scope.modules.delegates.onBind(scope.modules);
 			scope.modules.transactions.onBind(scope.modules);
@@ -249,10 +242,12 @@ describe('blocks/verify', function () {
 			delegates = scope.modules.delegates;
 			db = scope.db;
 			// Bus gets overwritten - waiting for mem_accounts has to be done manually
-			setTimeout(function () {
-				done();
-			}, 5000);
-		}, {db: db, bus: modulesLoader.scope.bus});
+			setTimeout(done, 5000);
+		});
+	});
+
+	after(function (done) {
+		application.cleanup(done);
 	});
 
 	function testValid (functionName) {
@@ -637,7 +632,7 @@ describe('blocks/verify', function () {
 				'forks_stat',
 				'votes where "transactionId" = \'17502993173215211070\''
 			], function (table, seriesCb) {
-				clearDatabaseTable(modulesLoader.db, modulesLoader.logger, table, seriesCb);
+				clearDatabaseTable(db, modulesLoader.logger, table, seriesCb);
 			}, function (err, result) {
 				if (err) {
 					return done(err);
